初步完成问题回复功能和错误信息回显

+++ src/main/java/motor/community/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
      * 编辑控制器
      */
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id, Model model) {
+    public String edit(@PathVariable(name = "id") Long id, Model model) {
         // 获取问题
         QuestionDTO question = questionService.getById(id);
         // 将问题相关信息传入Request域中并数据回显
@@ -62,7 +62,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
         if (title == null || title == "") {
